api/requirements.txt
# --- API Framework ---
# The main web framework for building the API.
fastapi==0.116.1

# --- ASGI Server ---
# The server that runs the FastAPI application.
# [standard] includes performance extras like websockets and httptools.
uvicorn[standard]==0.35.0

aiofiles==23.1.0

# --- Machine Learning ---
# TensorFlow Lite for edge inference (production-optimized)
tensorflow==2.18.0
# PIL for image processing
Pillow==10.4.0

# --- Core Dependencies ---
# NumPy for numerical computations (required by TensorFlow)
numpy>=1.21.0

# --- Development Dependencies (optional) ---
# Uncomment for development/testing:
# pytest>=7.0.0
# pytest-cov>=4.0.0
# keras>=2.12.0  # Only for preprocessing validation